fix: make logger backwards compatible
Several plugin use deleted logger methods in their hooks. Same is valid for `nativescript-cloud` extension.
To resolve this, get back the deleted methods, we'll delete them in 6.0.0 release.

Also, `nativescript-cloud` has an option `--workflow`, that is an object, while ours is boolean. This breaks its transpilation, so set ours to `any`.