fix: plugin create command needs more args for plugin seed
There are two new prompts in the plugin seed which require new command line options to be passed from CLI. Handle the new prompts and pass the new options. They can be used in automation scripts.
In order to prevent such issue in the future, ensure the spawned process's stdout and stdin are inherited, so if there's a prompt in the plugin seed, CLI will show it to the user.