fix: respect App_Resources directly from project
Currently App_Resources are copied to platforms folder and after that NativeScript CLI  moves them to the correct places in native project and deletes the App_Resources folder in platforms. On the other side, nativescript-dev-webpack plugin copies the App_Resources to platforms folder only on initial start (not when some file is changed). This led to the problem that when native file is changed, NativeScript CLI moves App_Resources to the correct places in native project and deletes App_Resources folder inside platforms (the one that nativescript-dev-webpack plugin has previously copied on initial webpack's compilation). On the next native file change, NativeScript CLI tries to move the platforms/App_Resources to the correct places inside native project, but this directory doesn't exist. So `cp unable to find source directory` error is thrown.

Fixes: https://github.com/NativeScript/nativescript-cli/issues/4377