fix: fix "InstallationLookupFailed" error when deploying on iOS device
"InstallationLookupFailed" error is thrown when trying to deploy on iOS device. The reason is that NativeScript CLI is not able to find applicationIdentifier from projectData.projectIdentifiers as the platform is not lower case.
For iOS simulator it is fine as the applicationIdentifier is not used when uploading files to the simulator.