fix: incorrect error for non-supported Node.js version
In case you are using unsupported Node.js version CLI must show you error that this version is not supported and to inform you which is the supported range. Currently this is not working and CLI shows error for `Unexpected token`.
The problem is that CLI has some specific logic to check the Node.js version and the code should be pure ES5 syntax in order to work with old Node.js versions (by default CLI's code is ES6).
The current issue is caused by using string interpolation in the mentioned code. Fix the message to fix the problem.