fix: unable to start iOS simulator when Android is not setup
In case CLI is used as a library, you need both Android and iOS to be setup in order to start simulator/emulator. The problem is that we check for errors in the setup before actually checking if we have found simulator/emulator with the specified name.
Change the logic and do not raise errors in case the specified name/id is found. This way you can start iOS Simulator without having any Android setup on local machine.