fix: restore application when the broken xml file is fixed
The files on android device are transferred through socket. When the application is in errorActivity, the runtime reports to the CLI that the application is in broken state (e.g. didRefresh property). In this case NativeScript CLI transfers all fallback files (bundle.js and runtime.js). CLI has a logic to ignore all non-existing files when transferring them on the device. As the fallback files are reported with relative paths, CLI consider them as non-existing and doesn't transfer them on device. So in order to fix this issue, we need to make fallbackFiles with absolute paths.

To reproduce:
1. tns run android
2. Broke some xml file from the application
3. Fix the broken xml file