fix: fix the workflow of preview-sync hook
The `kinvey-sdk` plugin relies on `preview-sync` hook and mainly on platform property from `hookArgs`. This PR changes the workflow of `preview-sync` hook and makes it the same as in NativeScript v5.4.0. In NativeScript v5.4.0 the hook is executed when pubnub reports the device that the QR code is scanned with. This way we're able to get the platform from device and adds it to the `hookArgs` passed to the hook.