fix: enable passing iOS codesigning options to LiveSync operations
Enable passing iOS Code signing options to operations like run/debug, etc. Currently the provision and teamId options are not passed to the controllers, so we do not use them during project preparation.
To resolve this, include the buildData in the deviceDescriptors - this way we'll have all the codesigning data required for the specific device and we'll also allow having different codesign options for different devices.
Also remove outputPath from deviceDescriptors as we can use it directly from the buildData.