fix: fix native change during livesync
Currently when a native file is changed during livesync on android, NativeScript CLI doesn't show the "Successfully synced application" message. This is due to the reason that NativeScript CLI tries to transfer native file on device (for example AndroidManifest.xml). After timeout of 30seconds, the socket on android device throws "Socket connection timeout" error but CLI doesn't show it as the parsing of received errors is not correct. This PR fixes the followings:
* the behavior when native file is changed during liveSync
* the parsing of errors during livesync
* stops livesync process when an error occurs during livesync