fix: hmr is working slowly
In recent changes we've broken HMR - in order to get its status on device, we need to call refresh application and monitor the logs. However, the logic at the moment is - wait for logs and call refresh application after that. This leads to very slow hmr. Also this way application will not be restarted when the HMR fails as we do not monitor the logs at the point when the hmr cannot be applied.