fix: fix prepare from plugin's directory
Currently when prepare command is executed from plugin's directory, CLI throws an error that it is not able to find `webpack.config.js` file. This is due to the reason that CLI finds nativescript key in package.json and consider that directory as project dir. This PR fixes this behavior as checks if there is `id` inside nativescript key.