fix: iOS device logs are not using source maps initially
When you run your app on iOS device for the first time, we execute full sync operation with a zip file. This does not gets in our logic for setting source maps of the transferred files and CLI does not use its logic to parse device logs and show the exact lines from original files.
Fix this by setting the source maps in this case.