fix: show the actual error when hook fails
When hook throws an error, there's no easy way to find exactly what is failing as CLI overwrites the error and its callstack. As it is not trivial solution to stop CLI of doing this, when a hook fails, just log the error with its callstack. We know these errors always lead to stopage of CLI process, so we can safely print them.