fix: execute npm install before all hooks
CLI must install all project dependencies before running any hook as the hooks are commonly created during the installation of packages. In the current codebase we install the depenencies only during `run` operation, however we actually need this to be executed before each project preparation.
To fix the behavior move the installation of dependencies in the prepareController and remove it from webpack-compiler-service. Also remove hook which is no longer supported - prepareJS.