fix: `tns debug ios` should work on device
Currently `tns debug ios` hangs when executed for device. The problem is that we pass `"waitForDebugger": false` to ios-device-lib binary, which can parse only strings. The value `false` is boolean, not string, so ios-device-lib binary crashes, but CLI does not understand this and hangs.
Fix this by ensuring we pass waitForDebugger as string.