fix: don't build every time the application from sidekick when using cloud build
Currently the application is built every time from sidekick when opening the app and starting cloud build. This happens as we decided that there are changes in app_resources as `prepareData.release` is undefined but `prepareInfo.release` is false. This way the check `prepareData.release !== this._prepareInfo.release` passes and we set `this._changesInfo.appResourcesChanged` to `true`.