fix: stop all webpack processes from sidekick when stopping the livesync
Currently when livesync is started on iOS device from sidekick, {N} CLI persist that livesync is started for iOS platform. After that if the livesync process is started on android device, {N} CLI overrides the persisted platforms and stores that the livesync is started for Android platform. However, when the livesync is stopped, {N} CLI stops it only for android platform. This means that the iOS webpack process is alive and continues to report data. This led to the multiple livesync messages when syncing the changes.