fix: stopPreview does not stop running webpack compilers
`stopPreview` method does not stop running webpack compilers, so it is no longer possible to run the same application on device (no matter preview, build or run). This happens in long living processes, like Sidekick.
To fix this, add new parameter to the method - an object containing the projectDir and use it to stop the processes.