fix: fix appstore list command for accounts with 2fa
Currently `tns appstore list` command throws an error when account is with 2fa and `--appleApplicationSpecificPassword` option is not provided. We shouldn't throw an error in this situation as we don't need `--appleApplicationSpecificPassword` on `tns appstore list` command as we don't use third party application. This option is mandatory only for `tns appstore upload` command where we use iTunesTransporter app when uploading the application.