fix: add missing await for isDynamicFramework check
There's a missing await in the check if an iOS Framework is dynamic. Linter rule (no-floating-promises) does not catch it, as it is not a pure call, i.e. linter will fail in case the code is:

```
this.isDynamicFramework(frameworkPath);
```

But it will not fail in case the method is called inside if statement (as it is in our code) or when you assing the value to a variable and do not await it anywhere:
```
const isDynamic = this.isDynamicFramework(frameworkPath);
```