fix: check correctly .framework files
The current check if `.framework` files are dynamically linked frameworks is not working as we pass only the framework name. This breaks all projects as tns-core-modules-widgets have `.framework`, CLI decides it is not dynamically linked and the application cannot run on device. Fix the check to pass the full path to the framework.