feat(publish): Enable publishing of iOS apps with Xcode 11
As `Application Loader` is no longer included in Xcode 11, CLI is not able to find the path to itmsTransporter and throws an error. However, itmsTransporter app is already part of ContentDeliveryServices.framework which is part of SharedFrameworks which are included in Xcode 11 itself. So this PR returns the correct path to itmsTransporter app based on xcode's version.