fix: handle changes in runtimes logs
In 6.1.0 there are several changes in the way both runtimes print logs. In order to handle this and keep the logs as they were with previous versions, introduce a check for the runtime version. In case the version is 6.1.0 or later (including prerelease versions) CLI will just replace the file reference in the message. For older versions CLI will still change the format of the message.

Add integration tests for the whole CLI logic for parsing the logs - the input of the tests is the real output from device and the tests verify CLI will print the expected messages to the output. In the resources of the test you can find .zip file containing the project used for running the tests. In case additional verifications are needed, unzip the content and add required tests.