feat: open feedback form when CLI is uninstalled
In case CLI is uninstalled, open a feedback form, so the user might explain why they've uninstalled the product.
Add analytics tracking for the uninstall call. Remove the `disableAnalytics` property of the `dev-preuninstall` command - this way we'll receive information for the number of times the command has been executed (usually it is called from CLI's preuninstall script).
Also add custom data in the analytics, so we'll be able to generate report to see how many times users uninstall the CLI (actually executing `npm uninstall...`) and if this happens from interactive terminal.