feat: add ipService to get current ip address
Add a new ipService to get the current ip address. As this information is critical for us, use three different ways to determine the ip address, so even if one of them fails, use the result from the next one.
Use the `play.nativescript.org/api/whoami` endpoint to get data for the public ip address.