fix: ensure data is sent to analytics on preuninstall
CLI's analytics process is a detached one, so during `preuninstall` command, CLI just sends information to the detached process what should be tracked and finishes CLI's execution. After that `npm` starts removing the CLI package and its `node_modules`.
At this point the analytics process may still work and trying to send data, but as it relies on components in node_modules, which npm currently deletes, it fails to send information to Google Analytics.
To ensure correct data is send to Analytics, ensure tracking is finished before ending the preuninstall command.