fix: save and remove of user-settings are not working
Due to recent changes, trying to save a setting in the user-settings.json or to remove it from there is not working. This breaks all new users and the execution of commands like `tns usage-reporting enable/disable`, `tns error-reporting enable/disable`.
The problem is an infinite recursion in the userSettingsService, which should just call jsonFileSettingsService, but instead it calls itself.
Fix the recursion and add unit tests for the userSettingsService to ensure it acts as a proxy to jsonFileSettingsService.