fix: list of apps is not working for iOS Devices
The installation_proxy that is used to list the applications on iOS Devices stops responding after 10-15 seconds and from now on CLI thinks the current application is not installed on device. This breaks `tns debug ios`, probably `tns run ios` as well.
Update to latest ios-device-lib where the installation_proxy is never cached.