fix: force refresh with socket when debugging the application
Currently there is an issue when livesyncing the application when it is stopped on breakpoint. The problem is that the application is always restarted - most probably due to issue with "RefreshRequest" notification. However, this notification was added as we needed to support syncing on wifi devices. As debug command doesn't work with wifi devices, we'll force refresh with socket on debug command.

Rel to: https://github.com/NativeScript/nativescript-cli/issues/4966