fix: pass appPath to the env from karma.conf.js
Two different webpack compilations are started on `tns test` command - the one from `karma-webpack` and the second one from `nativescript-dev-webpack`. As the karma's webpack doesn't provide appPath to the env and there is a hardcoded `appPath` inside `webpack.config.js` (to `src` for example), it works only with that hardcoded path.

Rel to: https://github.com/NativeScript/nativescript-cli/issues/5070