fix: pass allowProvisioningUpdates to xcodebuild only when building for device
The `-allowProvisioningUpdates` option is passed to `xcodebuild` no matter if building for device or for simulator.
According to the documentation of `xcodebuild`:
```
Allow xcodebuild to communicate with the Apple Developer website. For automatically signed targets, xcodebuild will create and update profiles, app IDs, and certificates. For manually signed targets, xcodebuild will download missing or updated provisioning profiles. Requires a developer account to have been added in Xcode's Accounts preference pane.
```

It seems it shouldn't be passed when building for simulator.