fix: persist the next hmr hash per platform
The NativeScript CLI persists the hash of the next hmr compilation locally and based on that decides if the reported hash is valid or not. However, in case when `tns run` command is executed, the application is randomly restarted on some devices. This happens as in this situation we have 2 webpack processes but the hmr hash is not persisted per platform.

Rel to: https://github.com/NativeScript/nativescript-cli/issues/5080