fix: values-v<sdk> directories are not prepared correctly
In Android's resources you can have different `values-v<sdk version>` directories which are used at runtime to differentiate resources for API levels.
In CLI we had a logic to include in native build only the `values-v<sdk version>` directories which are equal or lower to the SDK used for building the application.
This means that if you are using compileSdk 28, but you have `values-v29` directory in your resources it shouldn't be included in the native build.

Due to many changes in the structure of App_Resources/Android and the gradle template in NativeScript 4.0, this logic has been broken.
Fix it and execute it on every prepare of App_Resources.