fix: generate iOS images when we have their sizes
In case we do not have some of iOS Resource images in our image-definitions.json file, we do not generate images for them. This was not the case prior 6.2.0 release, so fix the behavior by getting back the old logic - get the image size from the Contents.json and generate image for it.
Also add the images from the templates that are currently missing from our `image-definitions.json` file.