fix: `tns info` should work with scoped modules
Currently the `tns info` command fails when executed inside application in which `tns-core-modules` package is not available as dependency. This may happen in case you use only `@nativescript/core` package.
The problem is that CLI is trying to read the package.json of the `<project dir>/node_modules/tns-core-modules` and this operation fails.
To fix this, add checks for both `tns-core-modules` and `@nativescript/core` packages. Warnings/information will be shown only for the package used in package.json, i.e. in case you use `tns-core-modules`, we'll show information only for it.