fix: emulator 29 is not started by default
In case you have emulator with API Level 29 and other older emulators, CLI should run the latest available (29), but it doesn't.
The problem is in the parsing of the string version - '9.0.0' is always newer than '10.0.0' which corresponds to API level 29. Fix the parsing, so now `tns run android` will correctly start latest available emulator in case there's nothing running.