fix: `tns update` command should work with npm tags
In the past `tns update` command worked with tags, but since 6.0 this was broken. The logic in Update Controller requires a valid package.json from template with same version. However, our templates do not have `next`, `rc`, etc. versions and passing tag to the command fails.
In order to fix this, construct the required template manually. For each of the packages we would update, check if the has a valid version with the same tag and get the current version behind this tag.