feat: improve detection of NativeScript plugins
Currently in case you have multiple occurences of a NativeScript plugin in node_modules, most of the time the build of application fails. For iOS Xcode fails with duplicate resources (most commonly frameworks) and for Android the Static Binding Generator will fail in case the JavaScript of a plugin differs.

Improve the handling by checking the versions of NativeScript plugins. In case the same version is installed multiple times, show warning to the user and link only one of the versions to the native build.
In case multiple versions are detected, throw an error - currently we cannot support this case.