fix: removed native code should be cleaned from native project
Currently in case you have native source code, it is added to the pbxproject's references. In case you remove the source code's directory, the files relations are not cleaned from the pbxproject, which makes the build fail. To fix this, instead of skipping the native code refrence for this folder, just call the method to add such pbxgroup and the nativescript-dev-xcode will cleanup the required resources.
Also fix the warning for modulemaps, it must be shown only when you have `.h` files and you do not have module.modulemap file at the root of the source code directory.