fix: ensure all temp files are cleaned
Currently CLI relies on temp module to create temporary files, which should be deleted once CLI process has no work left. However, in some cases, when Node.js crashes (or the process is killed with a signal) the temp module is unable to delete the files.
In order to fix this, introduce a tempService wrapper for temp module. For each path created by temp module, add a new cleanup action in the CLI's cleanup service, which will take care to delete the files even when the main CLI process dies with unusual error.