fix: file paths shown for device logs should be clickable
Currently CLI tries to persist the formating of logs as close as possible to the real logs coming from device.
The logs from the applications contain `file:///<filePath>`. CLI replaces the `<filePath>` with the real local file path. However, due to the `file:///` prefix, the link is not clickable. For example, in case you use VSCode's terminal, you should be able to use Ctrl + Click on the link and this will lead you to the exact line where the stack trace points to.
To resolve this issue, replace the `file:///` with just `file: `.