Allow passing command line arguments when running an app in the iOS Simulator
The `IOS_SIMULATOR_RUN_ARGS` environment variable allows specific command line arguments
to be passed to the simulator when running an app.

This is _extremely_ useful when working (for example) with Firebase and having to debug
events or other internals:

https://firebase.google.com/docs/analytics/debugview

The command line arguments `-FIRDebugEnabled` and `-FIRDebugDisabled` control the behavior
of Firebase in combination with its realtime event debugging console.

With this patch, enabling debugging is as easy as calling:

```bash
IOS_SIMULATOR_RUN_ARGS=-FIRDebugEnabled ns run ios
```